requirements.txt
streamlit>=1.24.0
pypdf>=3.9.0
faiss-cpu>=1.7.4
numpy>=1.24.3
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.3.1
pytest-cov>=4.1.0
python-magic>=0.4.27
tqdm>=4.65.0
langchain>=0.0.200